requirements.txt
Flask
python-binance
python-telegram-bot
python-dotenv
dateparser
pycryptodome
supabase